move compressImage to dedicated file
